refactor(donor-research): satisfy quality gate for the admin surface
- Extract listAdvisors/getAdminReport into donor-research-admin.service.ts so
  donor-research.service.ts stays under the 600-line budget (export the shared
  withMockSocialMetrics helper for the admin reader).
- memo() the AdvisorCard/ReportLink list items rendered in .map() loops.
- Refresh the stale quality baseline for the two metrics this PR increases
  (indexer.ts line count for the two new admin endpoint constants; react-doctor
  warning total). main had already drifted the indexer entry past baseline.
